Respect --disable-gpu command line option when terminal gpu acceleration set to auto
